Rename files *.reset to *.merger for attributeTypes
